requirements.txt
numpy
pillow
scipy